Testing with memcmp is probably not a good idea.
Depending on platform-dependent member sizes and layout, padding
may or may not be present, and in turn may or may not contain garbage.
We shouldn't `memcmp` with a literally-defined struct unless we're going
to first initialize it to a known value (padding and all), and even then
we should only do it when we know that our comparison will have been
similarly initialized. As the library function interfaces don't specify
how results will (or will not) be initialized, this sort of testing is
highly dependent on implementation detail and should be avoided.