More comprehensive checks in Pipeline builds.
The `tests` make target has a prerequisite on `lib`, but building that
way causes the library to be built with the more lenient compiler flags
usually reserved for tests. Explicitly building `lib` first causes it to
be built with the appropriate set of compiler flags.

We're also now building the examples which we weren't before.