Flexibility for name/version/build date defines.
Make it possible to override the build date in the same manner as the
library name and version: useful for making the build date match that of
a parent/including project.

Also add some sane fallbacks for the defines when they're unavailable
(e.g., when the library is being built as part of another project and
the defines are omitted).