Better detection of `-U` flag support for `ar(1)`.
Older versions of GNU ar(1) don't support the `-U` flag. In addition to
checking whether we're using GNU ar, we should also check its help
output to see whether it mentions the `-U` flag before trying to use it.
This is a bit fragile in that it will break if GNU ar ever changes the
format of its help output. However, the likelihood of that happening is
probably low, and our build would not break if it did: it would simply
cease to be incremental.