Use SVG logo for documentation.
Only slightly larger on disk, and much nicer on high-res displays.

Increased the native size of the image, too, from 55x55 to 80x80.
The Doxygen documentation for the `PROJECT_LOGO` option says that “the
maximum height of the logo should not exceed 55 pixels”, but in fact,
there's ugly whitespace above a 55px high image. The text to the right
of the image (project name + description) is 72px high (on my machine,
and my monitor, with my DPI, and with my fonts, and...), but happily
vertically centres to align with a larger image. So I'm going to assume
the note in the documentation applies to an older theme, and ignore it.